third_party/WebKit/LayoutTests/platform/win/fast/hidpi/image-srcset-intrinsic-size-zero-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
PASS imgWidth is 16777216
This test passes if the image is displayed with infinite dimensions.
